Add Swiss Bitcoin Pay as lightning payment processor
SBP will be used if phoenixd or lnd are not configured.

The implementation continuously polls the "payment" status in SBP, akin
to how Paypal, Stripe and Sum-up behaves. This is less than optimal
since SBP does provide a callback endpoint; unfortunatelly, there's
currently no mechanism to schedule or signal that a payment needs to be
updated asynchronously. This shall later be improved for all the
aforementioned payment processors.

The exchange rate for the cart is fixed at the moment the payment is
created; this is consistent with what is done for other lightning
payments. This means that if the price of the cart is in a fiat
currency, there will always be an inpredictable discrepancy between the
amount (in fiat) of the cart, the amount effectively paid (in sats) and
the result of the exchange as seen by the merchant in either SBP or
their bank account. This is because the actual exchange rate is not
used in the computation of the price. The situation could be improved
by providing the actual price of the cart to SBP and letting them fix
the amount in sats.

Since retrieving the state of a payment does not require authentication,
removing the SBP api-key does not affect existing payments; thus no
effort is done to stop polling those payments when the API key is
removed (thus payment are concluded normally).

The migration covers the case where the payment processor for a
lightning reciept (assumed to be lnd) is missing from the database.

The OpenAPI for the SBP API can be found in https://docs.swiss-bitcoin-pay.ch/

Signed-off-by: Roosembert Palacios <roos@be-bop.io>