🐛 mobile page displayed despite mobile version disabled #1897
Function `cmsFromContent` was augmented to support "forcing" the version
to display. If the requested version is available, it is displayed, else
the function defaults to the usual behaviour. What version should be
force is a policy enforced at the page level.

The landing page behaviour is special, conditionally displaying the
contents of the "home" page, if it exists. The query retrieving the data
for the page is slightly different than that of other pages; this query
was missing hasMobileContent and would thus ignore this setting.

By refactoring the logic on what page to choose, this issue was made
apparent. Thus, the query was sorted alphabetically tho make these kind
of issues more evident in the future.

Function `cmsFromContent`'s readability should be improved; in
particular, whether reactive content is available is a side-effect of
whether the array storing the mobile tokens is defined. This is a big
footgun, but shall be adressed in a different patch.

Because the arguments of `cmsFromContent` are now more explicit, some
renaming was done on other pages.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>