🦋 Add Tap-to-pay for automatic reconcilliation of POS external payments
The Tap-to-pay feature is accessible to POS operators in the Order
overview page. It is automatically enabled when the administrator
selects a provider for it; optionally, a configurable URL is opened when
the feature is activated.

When the POS operator taps 'Tap-to-pay', a process is launched for that
payment; before the process can start, it checks whether a conflicting
process exists; after the process is 'started', it checks again whether
a conflicting process exists (and bails-out if so).

The Tap-to-pay process is started by setting a specific property in the
order payment. The order updater (see `order-lock.ts`) processes all
pending payments; in particular point-of-sale payments with a Tap-to-pay
process. For this, it will try to reconciliate data from the selected
provider.

Since the Tap-to-pay process may be aborted out-of-band (e.g. the POS
operator closed the POS window), it is implemented so that its lifetime
is always known (ie. simulating a timeout); therefore, as Tap-to-pay
process may conclude without accepting a payment. Alternative
terminations include actually matching a payment or "cancelling" the
Tap-to-pay process; in both of these cases, the lifetime of the process
is updated to immediately expire the Tap-to-pay process; this makes the
"happy path" a simple optimization of the ongoing Tap-to-pay process.

A conflicting Tap-to-pay process is one whose lifetime overlaps with
that of another one. As an optimization, it is assumed time is "mostly"
monotonically increasing; this assumption may be violated if the system
time is reset. To account for potential clock skew, the payment event is
allowed to be within a few seconds of the Tap-to-pay process lifetime.
To prevent accidentally matching the same payment to multiple orders, an
extra delay is added when concluding the process in the "happy path".

The only currently implemented provider is Stripe.

Other changes:

- Added Spinner component.

- A constant containing all payment processors was added to enable
  parsing from a string. The type definition was changed to derive from
  this constant instead.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>