🐛 Fix price displayed when booking slots, some of which are free
The price computation was replicated in the frontend, for some reason.
The function computing the price in the backend is now exported and was
augmented to also include the price without discount.

Because we want the charged price to be as accurate as possible, the
price to bill and the price without discount are independently rounded.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>