🧹 Make code applying subscription discounts easier to read
Most of this code was extracted to a new function.

Incidentally, this fixes the following unexpected behaviours:

- Purchasing any number of subscriptions that two discounts grant some
  free amount of the same product would result in only one amount being
  applied.

- All discounts relative to the ensemble of all purchased subscriptions
  would be applied for each purchased subscription, leading to applying
  some unrelated discounts on some paid subscriptions.

NB: I have no clue why all scheduleEvents for an order containing a
subscription are confirmed when the order is paid. This is very
unexpected, but I don't want to break any possibly existing features, so
I'm leaving it as it was.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>