🐛 Fix account of free items for price computation of booking slots
The previous price computation assumed the number of free product units
was independent from between items in a cart. This is acceptable for as
long as no two-items refering to the same product are added to the card.

Since bookings (and booking slots) are the same product but are added as
different items, the assumption for the free items computation no-longer
holds.

Functions computing prices (and vat) for cart and items were refactored
to use an additional parameter specifying the number of free units (by
product) available to the user; the computation now correctly
sequentially deduces the free units even if multiple items with the same
product are specified.

The functions in src/lib/types/Cart.ts were moved to a separate file,
since they are unrelated to any 'Types' ; it seems they were placed
there by oversight.

Function `freeProductUnitsForCart` was introduced as a stop-gap since
the information about free products was previously retrieved along the
items. This is incorrect since the free product units are indexed by the
product and not by the item in the cart. Reduce the scope of the
changes, `freeProductUnitsForCart` is based in the information provided
by the cart items.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>