🐛 Do not use frontend values to compute free products
This commit removes the count of free units per item from the cart. This
was used to both compute the cart estimation and order price.

- Remove `freeQuantity` from cart items. This includes a database
  migration to removed any previously set values.
- Stop using the number of free product units provided by the frontend
  when computing cart and order prices.

Incidentally, this also fixes the computation of the number of free
items for booking slots, since we use the same computation as the order
placement code.

The `cart` property of the root layour was updated to reflect properties
of the cart (and not the cart items).

Signed-off-by: Roosembert Palacios <roos@be-bop.io>