🦋 Display refusal reason when processing card payments
This commit displays the refusal reason to the user for Card payments
with for Stripe, Sumsub or Paypay.

If the payment processor does not directly address a message for the
user, but provides an advisory, a message is constructed containing
that advisory.

This commit also fixes some unrelated translation strings.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>