Enable loading environment variables at runtime
Environment variables would previously be inlined at build-time. This
is problematic because it prevents us from shipping a release since
secrets such as the database URL are, in general, not available at
build-time. In an effort to keep this simple, most environment
variables were moved into a file and the loading has been modified as
follows:

- The environment variables available at build-time are baked as default
  values in the build.
- If the environment variable is present at runtime, the build-time
  value is replaced.

Some “control” environment variables (such as PUBLIC_VERSION, VITEST)
were opportunistically left as-is, because the information is obviously
available at build-time. No effort was made to determine whether each
variable is already available at build-time.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>