be-BOP wizard Debian package 🚀
This commit adds:

- A shell script to manage be-BOP installations.
- A configurable debian package to install and configure the system.

Because of how Debian package transactions work, the setup process is
currently two-fold: First, install the debian package; then run the
wizard to finish configuration. This may be improved in the future, but
it's the simplest solution. This is because we cannot change the
packages to install once the installation process has started.

The wizard provides a single option, that is to “update” the be-BOP
release. Further improvements may be added in the future (e.g.
automatically download new versions daily).

Signed-off-by: Roosembert Palacios <roos@be-bop.io>