Recognise SSO emails as the user secondary emails
When a user connects via SSO, the user may end-up linking multiple
identities. Each identity may provide us with an email address.

This commit extends the UserIdentifier to contain these secondary
emails. This is useful, for example when searching the discounts
belonging to a user.

This that this commit accidentally persists the secondary emails. This
is because there is no distinction between the domain object
UserIdentifier and the database object UserIdentifier. Unfortunatelly,
there is no time available to split these. This is also part of the
reason why there are now `email` and `secondaryEmails` in UserIdentifier.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>