Initial implementation of order tabs
An order tab holds products and quantities, that will be added into a
cart and an order at a later time.

At this time, tabs can only be accessed by users with the POS role. A
tab is independent of a user. A tab contains lines, each line contains a
product and a quantity.

A collection of lines may be used to create a cart, in which case, the
lines learn the cart they're associated to. When the cart is
checked-out, the lines also learn the order they're associated to. The
current implementation removes the orderTab once all the orders have
been paid.

The current implementation handles gracefully when a cart previously
associated to a tab is removed; likewise when an order associated to a
tab is cancelled or expired.

Because tabs are accessible by any POS user, a cart may be associated to
a tab, but then abandoned by the user (e.g. employee clocks out); in
which case, the associated lines cannot be checked-out. This remains to
be specified.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>