POS/touch: include tab slug in URL path
The order tab was previously specified as a URL query parameter. Thus,
when switching from one tab to the next one, we'd have to manually
invalidate all load dependencies. This causes performance issues because
the products do not change from one tab to the other one.

This commit includes the tab slug in the URL, so the usual invalidation
rules apply. This enables us change the tab without invalidation the
products (and other static data).

Signed-off-by: Roosembert Palacios <roos@be-bop.io>