Record order tab slug along with cart and order
This enables to search at a later time what order tab is related to what
cart. We can also display a button on the order interface to go to the
order tab.

The update of what elements are paid has also been modified to use this
information to restrict the documents to update (previously all
documents would be updated, even if the update did not perform any
modifications).

Signed-off-by: Roosembert Palacios <roos@be-bop.io>