CMS: Support pages that bypass the application layout
Introduce the `$page.data.suppressAppLayout` flag to allow certain CMS-driven
pages to render full HTML documents (including <head> and <body>) directly.
When this flag is set, the layout injects the provided head into <svelte:head>
and renders the body content as-is, skipping all app-level wrappers and styles.

This flag is only set by **some** CMS pages with the “advanced HTML” mode.
Specifically, the flag is set if the page contains an HTML5 document (ie.
a document starting with a `<!doctype html>` tag). This enables the “advanced
HTML” CMS mode to serve static-like documents while preserving normal Svelte
rendering for standard pages aswell as previous CMS pages with the
“advanced HTML” mode that relied in the hosted behaviour.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>